fix encoding issue, replace long with int, fix range

+++ plugin/parsers/microdvd.py
@@ -77,8 +77,8 @@ class MicroDVDParser(BaseParser):
 
         for m in re.finditer("\{(\d+)\}\{(\d+)\}(.*)", text):
             try:
-                startTime = float(long(m.group(1)) / float(fps)) * 1000
-                endTime = float(long(m.group(2)) / float(fps)) * 1000
+                startTime = float(int(m.group(1)) / float(fps)) * 1000
+                endTime = float(int(m.group(2)) / float(fps)) * 1000
                 text = '\n'.join(m.group(3).split('|'))
                 subs.append(self.createSub(text, startTime, endTime))
             except Exception as e:
